josh-talks-asr/requirements.txt
# Josh Talks — Hindi ASR Pipeline
# Core ML / Audio
torch>=2.2.0
torchaudio>=2.2.0
transformers>=4.38.0
datasets>=2.18.0
evaluate>=0.4.0

# LoRA + 8-bit Quantization
peft>=0.10.0
bitsandbytes>=0.43.0

# Language detection
fasttext-wheel>=0.9.2

# Spell checking
symspellpy>=6.7.7

# WER evaluation
editdistance>=0.8.1
jiwer>=3.0.3

# Hindi NLP
indic-nlp-library>=0.91

# Data / Utils
numpy>=1.26.0
pandas>=2.2.0
requests>=2.31.0
